Move fixedpoint back to its own file
