+++ examples/kaplanmeier.py
@@ -32,7 +32,8 @@ from scipy.stats import chi2
 
 import pysnark.runtime
 
-from pysnark.runtime import PrivVal, LinCombFxp
+from pysnark.runtime import PrivVal
+from pysnark.fixedpoint import LinCombFxp
 
 if __name__ == '__main__':
     # example batch of survival data from two populations
